+++ packages/contact-manager/__tests__/shared/contactManagerAgentLogic.ts
@@ -4,7 +4,8 @@ import {
   ElectronicAddress,
   GetPartiesArgs,
   Identity,
-  IdentityRole,
+  CredentialRole,
+  IdentityOrigin,
   NaturalPerson,
   NonPersistedElectronicAddress,
   NonPersistedIdentity,
@@ -45,7 +46,8 @@ export default (testContext: { getAgent: () => ConfiguredAgent; setup: () => Pro
       const correlationId = 'default_example_did'
       const identity: NonPersistedIdentity = {
         alias: correlationId,
-        roles: [IdentityRole.ISSUER, IdentityRole.VERIFIER],
+        origin: IdentityOrigin.EXTERNAL,
+        roles: [CredentialRole.ISSUER, CredentialRole.VERIFIER],
         identifier: {
           type: CorrelationIdentifierType.DID,
           correlationId,
@@ -241,7 +243,8 @@ export default (testContext: { getAgent: () => ConfiguredAgent; setup: () => Pro
       const correlationId = 'new_example_did'
       const identity: NonPersistedIdentity = {
         alias: correlationId,
-        roles: [IdentityRole.ISSUER, IdentityRole.VERIFIER],
+        origin: IdentityOrigin.EXTERNAL,
+        roles: [CredentialRole.ISSUER, CredentialRole.VERIFIER],
         identifier: {
           type: CorrelationIdentifierType.DID,
           correlationId,
@@ -265,7 +268,8 @@ export default (testContext: { getAgent: () => ConfiguredAgent; setup: () => Pro
       const correlationId = 'missing_connection_add_example'
       const identity: NonPersistedIdentity = {
         alias: correlationId,
-        roles: [IdentityRole.ISSUER, IdentityRole.VERIFIER],
+        origin: IdentityOrigin.EXTERNAL,
+        roles: [CredentialRole.ISSUER, CredentialRole.VERIFIER],
         identifier: {
           type: CorrelationIdentifierType.URL,
           correlationId,
@@ -281,7 +285,8 @@ export default (testContext: { getAgent: () => ConfiguredAgent; setup: () => Pro
       const correlationId = 'missing_connection_update_example'
       const identity: NonPersistedIdentity = {
         alias: correlationId,
-        roles: [IdentityRole.ISSUER, IdentityRole.VERIFIER],
+        origin: IdentityOrigin.EXTERNAL,
+        roles: [CredentialRole.ISSUER, CredentialRole.VERIFIER],
         identifier: {
           type: CorrelationIdentifierType.DID,
           correlationId,
@@ -297,7 +302,8 @@ export default (testContext: { getAgent: () => ConfiguredAgent; setup: () => Pro
       const correlationId = 'new_update_example_did'
       const identity: NonPersistedIdentity = {
         alias: 'update_example_did',
-        roles: [IdentityRole.ISSUER, IdentityRole.VERIFIER],
+        origin: IdentityOrigin.EXTERNAL,
+        roles: [CredentialRole.ISSUER, CredentialRole.VERIFIER],
         identifier: {
           type: CorrelationIdentifierType.DID,
           correlationId: 'update_example_did',