+++ packages/event-logger/src/agent/EventLogger.ts
@@ -1,8 +1,9 @@
-import { IAgentPlugin } from '@veramo/core'
 import { AbstractEventLoggerStore } from '@sphereon/ssi-sdk.data-store'
-import { AuditLoggingEvent, LoggingEventType, LogLevel } from '@sphereon/ssi-sdk.core'
+import { IAgentPlugin } from '@veramo/core'
+import { Loggers, LoggingEventType, LogLevel, LogMethod } from '@sphereon/ssi-types'
+import { AuditLoggingEvent } from '@sphereon/ssi-sdk.core'
 import { v4 as uuidv4 } from 'uuid'
-import { schema } from '../index'
+import { NonPersistedAuditLoggingEvent, schema } from '../index'
 import { EventLoggerOptions, GetAuditEventsArgs, IEventLogger, RequiredContext, LogAuditEventArgs, LoggingEvent } from '../types/IEventLogger'
 
 /**
@@ -10,24 +11,42 @@ import { EventLoggerOptions, GetAuditEventsArgs, IEventLogger, RequiredContext, 
  */
 
 // Exposing the methods here for any REST implementation
-export const eventLoggerAuditMethods: Array<string> = ['loggerGetAuditEvents', 'loggerLogAuditEvent']
+export const eventLoggerAuditMethods: Array<string> = ['loggerGetAuditEvents', 'loggerLogAuditEvent', 'loggerLogGeneralEvent']
 export const eventLoggerMethods: Array<string> = [...eventLoggerAuditMethods]
 
 export class EventLogger implements IAgentPlugin {
   readonly schema = schema.IEventLogger
   readonly eventTypes: Array<LoggingEventType> = []
+  private readonly store?: AbstractEventLoggerStore
+  readonly simpleLoggers: Loggers
 
   readonly methods: IEventLogger = {
     loggerGetAuditEvents: this.loggerGetAuditEvents.bind(this),
     loggerLogAuditEvent: this.loggerLogAuditEvent.bind(this),
+    loggerLogGeneralEvent: this.loggerLogGeneralEvent.bind(this),
   }
-
-  private readonly store?: AbstractEventLoggerStore
 
   constructor(options: EventLoggerOptions) {
     const { store, eventTypes } = options
+    const generalOpts = options.general ?? { debugPkg: true }
     this.store = store
     this.eventTypes = eventTypes
+
+    const methods: Array<LogMethod> = []
+    if (generalOpts.debugPkg) {
+      methods.push(LogMethod.DEBUG_PKG)
+    }
+    if (generalOpts.console) {
+      methods.push(LogMethod.CONSOLE)
+    }
+    if (generalOpts.events) {
+      methods.push(LogMethod.EVENT)
+    }
+    this.simpleLoggers = new Loggers({
+      methods,
+      eventName: generalOpts.eventName,
+      defaultLogLevel: generalOpts.defaultLogLevel,
+    })
   }
 
   public async onEvent(event: LoggingEvent, context: RequiredContext): Promise<void> {
@@ -35,6 +54,11 @@ export class EventLogger implements IAgentPlugin {
       case LoggingEventType.AUDIT:
         // Calling the context of the agent to make sure the REST client is called when configured
         await context.agent.loggerLogAuditEvent({ event: event.data })
+        break
+      case LoggingEventType.GENERAL:
+        // Calling the context of the agent to make sure the REST client is called when configured
+        // TODO: We might also want to do this locally though, as these logs are not persisted typically
+        await context.agent.loggerLogGeneralEvent({ event: event.data })
         break
       default:
         return Promise.reject(Error(`Event type ${event.type} not supported`))
@@ -49,6 +73,12 @@ export class EventLogger implements IAgentPlugin {
     }
 
     return this.store.getAuditEvents({ filter })
+  }
+
+  private async loggerLogGeneralEvent(args: LogAuditEventArgs): Promise<NonPersistedAuditLoggingEvent> {
+    const { event } = args
+    this.simpleLoggers.get(event.data.system).logl(event.data.level ?? LogLevel.INFO, event.data.data, event.data)
+    return args.event
   }
 
   private async loggerLogAuditEvent(args: LogAuditEventArgs): Promise<AuditLoggingEvent> {