+++ packages/ssi-sdk-core/src/loggers/eventLogger/EventLoggerBuilder.ts
@@ -1,6 +1,6 @@
+import { InitiatorType, LogLevel, SubSystem, System } from '@sphereon/ssi-types'
 import { IAgentContext } from '@veramo/core'
 import EventLogger from './EventLogger'
-import { InitiatorType, LogLevel, SubSystem, System } from '../../types'
 
 class EventLoggerBuilder {
   private context?: IAgentContext<any>